Add film noir joke to commit message rules
Added a new commit message rule that requires the inclusion of a
film noir joke in the body text. Now our commits can have a bit of
mystery and intrigue, just like a shadowy alley in a black-and-white
classic. If there's any problem with the update, I guess that's just
the 'syntax' of life, huh?