Update git-hooks commit message rule
Updated the rule set in the 'git-hooks/fix-commit-message' script. The change
is a switch from requiring a film noir joke in the body of the commit message
to requiring a Sam Altman joke. Speaking of which, why did Sam Altman go to
Mars? To check if the Martians had start-ups, too!